Game Works!
The game now has a basic UI, Gameplay structure, and a simple AI with targeting. I have encountered an error with the target mode. Also updated the random move generator to cycle over the board and find empty spots, add them to a list, and randomly generate an index for that list. This will avoid issues further into gameplay